reconfigure the sleep timer stuff so that it does not run in native anymore
